Added basic Date factory methods

+++ src/test/java/net/xmercerweiss/mdc/tests/ModernDigitalChronologyTest.java
@@ -12,57 +12,57 @@ import net.xmercerweiss.mdc.*;
 
 public class ModernDigitalChronologyTest
 {
-  private static final Map<Long,Long> PROLEPTIC_TO_FIRST_EPOCH_DAY = Map.ofEntries(
-    entry(-100L, -36524L),
-    entry(-99L, -36159L),
-    entry(-98L, -35794L),
-    entry(-70L, -25567L),
-    entry(-22L, -8036L),
-    entry(-18L, -6575L),
-    entry(-14L, -5114L),
-    entry(-11L, -4018L),
-    entry(-10L, -3653L),
-    entry(-9L, -3287L),
-    entry(-8L, -2922L),
-    entry(-7L, -2557L),
-    entry(-6L, -2192L),
-    entry(-5L, -1826L),
-    entry(-4L, -1461L),
-    entry(-3L, -1096L),
-    entry(-2L, -731L),
-    entry(-1L, -365L),
-    entry(0L, 0L),
-    entry(1L, 365L),
-    entry(2L, 730L),
-    entry(3L, 1096L),
-    entry(4L, 1461L),
-    entry(5L, 1826L),
-    entry(6L, 2191L),
-    entry(7L, 2557L),
-    entry(8L, 2922L),
-    entry(9L, 3287L),
-    entry(10L, 3652L),
-    entry(11L, 4018L),
-    entry(14L, 5113L),
-    entry(15L, 5479L),
-    entry(18L, 6574L),
-    entry(19L, 6940L),
-    entry(22L, 8035L),
-    entry(23L, 8401L),
-    entry(30L, 10957L),
-    entry(31L, 11323L),
-    entry(98L, 35794L),
-    entry(99L, 36160L),
-    entry(100L, 36525L)
+  private static final Map<Integer,Long> PROLEPTIC_TO_FIRST_EPOCH_DAY = Map.ofEntries(
+    entry(-100, -36524L),
+    entry(-99, -36159L),
+    entry(-98, -35794L),
+    entry(-70, -25567L),
+    entry(-22, -8036L),
+    entry(-18, -6575L),
+    entry(-14, -5114L),
+    entry(-11, -4018L),
+    entry(-10, -3653L),
+    entry(-9, -3287L),
+    entry(-8, -2922L),
+    entry(-7, -2557L),
+    entry(-6, -2192L),
+    entry(-5, -1826L),
+    entry(-4, -1461L),
+    entry(-3, -1096L),
+    entry(-2, -731L),
+    entry(-1, -365L),
+    entry(0, 0L),
+    entry(1, 365L),
+    entry(2, 730L),
+    entry(3, 1096L),
+    entry(4, 1461L),
+    entry(5, 1826L),
+    entry(6, 2191L),
+    entry(7, 2557L),
+    entry(8, 2922L),
+    entry(9, 3287L),
+    entry(10, 3652L),
+    entry(11, 4018L),
+    entry(14, 5113L),
+    entry(15, 5479L),
+    entry(18, 6574L),
+    entry(19, 6940L),
+    entry(22, 8035L),
+    entry(23, 8401L),
+    entry(30, 10957L),
+    entry(31, 11323L),
+    entry(98, 35794L),
+    entry(99, 36160L),
+    entry(100, 36525L)
   );
 
   @Test
   void FirstEpochDayOfYear_WithProleptic_ProducesExpected()
   {
     ModernDigitalChronology chrono = ModernDigitalChronology.INSTANCE;
-    for (Entry<Long,Long> e : PROLEPTIC_TO_FIRST_EPOCH_DAY.entrySet())
+    for (Entry<Integer,Long> e : PROLEPTIC_TO_FIRST_EPOCH_DAY.entrySet())
     {
-      long proleptic = e.getKey();
+      int proleptic = e.getKey();
       long expected = e.getValue();
       long actual = chrono.epochDay(
         proleptic,  // Year
